i32 kernel native mode (and tests fixes)

+++ linalg/src/generic/mmm.rs
@@ -173,7 +173,7 @@ where
                     if TI::datum_type().is_float() {
                         add_mat_mul::<MR, NR, TI, TI, TI>(pa, pb, k, &mut ab);
                     } else {
-                        add_mat_mul::<MR, NR, i32, i8, i8>(pa, pb, k, std::mem::transmute(&mut ab))
+                        add_mat_mul::<MR, NR, i32, i32, i32>(pa, pb, k, std::mem::transmute(&mut ab))
                     }
                     /*
                      */
